pyproject.toml
[project]
name = "predictor-cancer-rectal"
version = "0.1.0"
description = "Proyecto academico para clasificacion de polipos y patologias relacionadas con cancer rectal."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "pillow>=11.0.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.0",
    "seaborn>=0.13.2",
    "streamlit>=1.44.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]
